refactor(web): atomize photo loader routes
