+++ src/web/photo_loader/__init__.py
@@ -1,0 +1,2 @@
+"""Internal layers for the photo loader web routes."""
+
